decode b64 result back to unicode before str interp
In #440, we learn the reddit authorization string that composes key and
secret into a base-64 string leaves it as a bytestring, and then tries
in python3 tries to insert it into a unicode literal.  That adds its
repr format "b'foo'" in the auth, which is wrong.

Instead encode the base64 string back to unicode. The b64 format is
guaranteed to be 7-bit safe, so even lame default encodings should
handle it fine.